. appmon/Makefile.am, appfuncs.sh, initappreg.sh: added initappregsh, appfuncs.sh for intiailization of application registry. . file.cpp, regfs.cpp: fixed two bugs. One is the gid was mistakenly using guint16 when setting gid, and caused gid greater than 64k is truncated.And the other was failure to handle root directory correctly in Chown and Chmod. . pubfuncs.sh, rpcfshow.sh: added 'list users' and 'list groups' to rpcfshow.sh. . index_cn.md: update the url to 'how to build rpc-frmwrk'.

+++ monitor/registry/file.cpp
@@ -1522,7 +1522,7 @@ void CFileImage::SetGid( gid_t wGid )
     gint32 ret = 0;
     do{
         WRITE_LOCK( this );
-        m_oInodeStore.m_wgid = ( guint16 )wGid;
+        m_oInodeStore.m_wgid = wGid;
         UpdateCtime();
         ret = this->Flush( FLAG_FLUSH_INODE );
         if( ERROR( ret ) )
@@ -1535,7 +1535,7 @@ void CFileImage::SetUid( uid_t wUid )
     gint32 ret = 0;
     do{
         WRITE_LOCK( this );
-        m_oInodeStore.m_wuid = ( guint16 )wUid;
+        m_oInodeStore.m_wuid = wUid;
         UpdateCtime();
         ret = this->Flush( FLAG_FLUSH_INODE );
         if( ERROR( ret ) )
